Slightly improve helper design and use RunCommand
For the helpers, the previous style was shorter but looked off (an array with 0 elements),
so switch to a better one. toArray() with no arguments requires a cast
which reduces type safety.

According to wpilib docs, .setDefaultCommand() should take a command
that does not finish, so we should use RunCommand rather than
InstantCommand, even though both have the same effect.